+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # MountsJournal
 
 * Updated the Selected filter.
+* Added macro buttons to the journal.
+* Fixed pickup mounts in combat.